+++ tests/test_auth.py
@@ -25,7 +25,7 @@ def auth_enabled_client():
         mock_settings.browser_headless = True
         mock_settings.message_enabled = False
         mock_settings.auth_enabled = True
-        mock_settings.zitadel_issuer = "http://localhost:8085"
+        mock_settings.yggdrasil_issuer = "http://localhost:8085"
         mock_settings.jwt_audience = ""
 
         from fenrir.main import create_app
@@ -48,7 +48,7 @@ def auth_disabled_client():
         mock_settings.browser_headless = True
         mock_settings.message_enabled = False
         mock_settings.auth_enabled = False
-        mock_settings.zitadel_issuer = ""
+        mock_settings.yggdrasil_issuer = ""
         mock_settings.jwt_audience = ""
 
         from fenrir.main import create_app